Fix missing keys when trimming a mot animation
Some joints may rely on linear interpolation heavy. When cutting an animation with large gaps between keys, the user may choose a range for which there are no keys. To handle this we need to invent new linearly interpolated keys to stand at the ends of the list of keys so that information is not lost.